RWA-2634: Detecting Replication Problems
